% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPanelcnMops.R
\docType{data}
\name{control}
\alias{control}
\title{GRanges object of countWindows with read counts for control samples 
as elementMetadata}
\description{
Control data included in panelcn.mops
}
\author{
Gundula Povysil
}
\keyword{data}
