% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPanelcnMops.R
\docType{data}
\name{countWindows}
\alias{countWindows}
\title{result object of getWindows - a data.frame with the contents of 
the provided BED file with an additional gene name and exon name column}
\description{
Data included in panelcn.mops
}
\author{
Gundula Povysil
}
\keyword{data}
