% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPanelcnMops.R
\docType{data}
\name{read.width}
\alias{read.width}
\title{read width used for calculating RCs of test and control}
\description{
Data included in panelcn.mops
}
\examples{
data(panelcn.mops)
read.width
}
\author{
Gundula Povysil
}
\keyword{data}
