% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPanelcnMops.R
\docType{data}
\name{resultlist}
\alias{resultlist}
\title{result object of runPanlecnMops - a list of instances of 
"CNVDetectionResult"}
\description{
Result data included in panelcn.mops
}
\author{
Gundula Povysil
}
\keyword{data}

