% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPanelcnMops.R
\docType{data}
\name{control}
\alias{control}
\title{GRanges object of countWindows with read counts for control samples 
as elementMetadata.}
\description{
The object was created using the function 
countBamListInGRanges with the enclosed countWindows object, a subset of 
BAM files provided by the 1000 Genomes Project and the read.width parameter 
set to 150.
}
\details{
Control data included in panelcn.mops
}
\examples{
data(panelcn.mops)
control
}
\author{
Gundula Povysil
}
\keyword{data}
