% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPanelcnMops.R
\docType{data}
\name{test}
\alias{test}
\title{GRanges object of countWindows with read counts for a test sample as 
elementMetadata}
\description{
Test data included in panelcn.mops
}
\author{
Gundula Povysil
}
\keyword{data}
